# Parallelize proposal transactions signature verification