--- a/ledger/eval.go
+++ b/ledger/eval.go
@@ -941,7 +941,7 @@ func (validator *evalTxValidator) run() {
 		CurrProto: validator.block.BlockHeader.CurrentProtocol,
 	}
 	var unverifiedTxnGroups [][]transactions.SignedTxn
-	unverifiedTxnGroups = make([][]transactions.SignedTxn, len(validator.txgroups))
+	unverifiedTxnGroups = make([][]transactions.SignedTxn, 0, len(validator.txgroups))
 	for _, group := range validator.txgroups {
 		signedTxnGroup := make([]transactions.SignedTxn, len(group))
 		for j, txn := range group {
@@ -1027,6 +1027,8 @@ func eval(ctx context.Context, l ledgerForEvaluator, blk bookkeeping.Block, vali
 	if validate {
 		// wait for the validation to complete.
 		select {
+		case <-ctx.Done():
+			return StateDelta{}, ctx.Err()
 		case err, open := <-txvalidator.done:
 			if !open {
 				break