## Summary

Our existing code was verifying the transactions in parallel with the insertion to the evaluator, but that won't fully utilize the host computational abilities. To push it to the end, we need to break the transaction verification into subsets and parallelize the verification.

This is the first PR, which would handle the verification parallelization. The subsequent PR would ensure that we don't revalidate the same transaction coming from two different proposals, regardless if the transaction appear in the transaction pool or not.

## Test Plan

Added unit test to verify changed code correctness.

## Performance

Testing the performance gains between running `verify.TxnPool` vs using the new `verify.PaysetGroups`:
| Transaction Count | Before ( serialized ) | After ( parallelized ) |
|:--:|:--:|:--:|
|1 (amortized) |65,200 ns| 14,600 ns|
|5,000|326 ms|73 ms|
|10,000|656 ms|146 ms|
|50,000|3.260 s|730 ms|

Tests were conducted on a MacBook Pro 2.7 GHz Quad-Core Intel Core i7