@algorandskiy 2020-12-09
I'd love to see detailed explanation in the commit message like "split entire list of txn groups to X blocks and post them to processing in parallel".